DjangoTaskV1/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.5
django-dirtyfields==1.9.6
django-model-utils==5.0.0
django-notifications-hq==1.8.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
jsonfield==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
packaging==24.2
pillow==11.1.0
pluggy==1.6.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
PyPDF2==3.0.1
pytest==8.4.1
pytest-django==4.11.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
